chore: move new sql into pg-store-v2 class
